Move PostBuild to reflex.
