+++ src/Reflex/Dom/Old.hs
@@ -68,7 +68,7 @@ import Reflex.Dom.Builder.Class
 import Reflex.Dom.Builder.Immediate
 import Reflex.Dom.Internal.Foreign
 import Reflex.PerformEvent.Class
-import Reflex.Dom.PostBuild.Class
+import Reflex.PostBuild.Class
 import Reflex.Dom.Widget.Basic
 import Reflex.Host.Class
