+++ src/Reflex/Dom/Widget/Input.hs
@@ -44,7 +44,7 @@ import Reflex.Dom.Builder.Class
 import Reflex.Dom.Builder.Immediate
 import Reflex.Dom.Class
 import Reflex.PerformEvent.Class
-import Reflex.Dom.PostBuild.Class
+import Reflex.PostBuild.Class
 import Reflex.Dom.Widget.Basic
 import qualified Text.Read as T
